refactor: move point clouds stuff into its own module
