+++ viewer/core/src/public/migration/Cognite3DViewer.ts
@@ -7,8 +7,6 @@ import TWEEN from '@tweenjs/tween.js';
 import omit from 'lodash/omit';
 import { Subscription, fromEventPattern } from 'rxjs';
 
-import { LoadingState } from '@reveal/cad-geometry-loaders';
-
 import { defaultRenderOptions, SsaoParameters, SsaoSampleQuality, AntiAliasingMode } from '@reveal/rendering';
 
 import {
@@ -21,7 +19,13 @@ import {
 } from '@reveal/utilities';
 
 import { MetricsLogger } from '@reveal/metrics';
-import { intersectCadNodes, CadModelSectorLoadStatistics } from '@reveal/cad-model';
+import { intersectCadNodes, CadModelSectorLoadStatistics, Cognite3DModel } from '@reveal/cad-model';
+import {
+  intersectPointClouds,
+  PointCloudIntersection,
+  PointCloudBudget,
+  CognitePointCloudModel
+} from '@reveal/pointclouds';
 
 import {
   AddModelOptions,
@@ -29,20 +33,16 @@ import {
   Intersection,
   CameraChangeDelegate,
   PointerEventDelegate,
-  CadModelBudget,
-  PointCloudBudget
+  CadModelBudget
 } from './types';
 import { NotSupportedInMigrationWrapperError } from './NotSupportedInMigrationWrapperError';
 import RenderController from './RenderController';
-import { Cognite3DModel } from '@reveal/cad-model';
-import { CognitePointCloudModel } from './CognitePointCloudModel';
 import { RevealManager } from '../RevealManager';
 import { DisposedDelegate, SceneRenderedDelegate, RevealOptions } from '../types';
 
 import { Spinner } from '../../utilities/Spinner';
-import { intersectPointClouds } from '../../datamodels/pointcloud/picking';
 
-import { CadIntersection, IntersectionFromPixelOptions, PointCloudIntersection } from '../..';
+import { CadIntersection, IntersectionFromPixelOptions } from '../..';
 import { PropType } from '../../utilities/reflection';
 import { ViewerState, ViewStateHelper } from '../../utilities/ViewStateHelper';
 import { RevealManagerHelper } from '../../storage/RevealManagerHelper';
@@ -50,7 +50,7 @@ import { RevealManagerHelper } from '../../storage/RevealManagerHelper';
 import { DefaultCameraManager, CameraManager } from '@reveal/camera-manager';
 import { CdfModelIdentifier, File3dFormat } from '@reveal/modeldata-api';
 import { DataSource, CdfDataSource, LocalDataSource } from '@reveal/data-source';
-import { IntersectInput, SupportedModelTypes, CogniteModelBase } from '@reveal/model-base';
+import { IntersectInput, SupportedModelTypes, CogniteModelBase, LoadingState } from '@reveal/model-base';
 
 import { CogniteClient } from '@cognite/sdk';
 import log from '@reveal/logger';
