+++ viewer/packages/cad-parsers/index.ts
@@ -3,7 +3,6 @@
  */
 
 export { CadMetadataParser } from './src/metadata/CadMetadataParser';
-export { MetadataRepository } from './src/metadata/MetadataRepository';
 export { CadModelMetadataRepository } from './src/metadata/CadModelMetadataRepository';
 export { CadModelMetadata } from './src/metadata/CadModelMetadata';
 export { SectorSceneImpl } from './src/utilities/SectorScene';
