+++ viewer/packages/model-base/index.ts
@@ -2,8 +2,9 @@
  * Copyright 2022 Cognite AS
  */
 
-export { IntersectInput } from './src/types';
+export { IntersectInput, LoadingState } from './src/types';
 export { SupportedModelTypes } from './src/SupportedModelTypes';
 export { CogniteModelBase } from './src/CogniteModelBase';
 
 export { NodeIdAndTreeIndexMaps } from './src/NodeIdAndTreeIndexMaps';
+export { MetadataRepository } from './src/MetadataRepository';
