+++ viewer/packages/model-base/package.json
@@ -8,6 +8,7 @@
     "lint": "yarn ws:lint"
   },
   "dependencies": {
+    "@reveal/modeldata-api": "workspace:*",
     "@reveal/nodes-api": "workspace:*",
     "@reveal/utilities": "workspace:*"
   }
