+++ viewer/packages/pointclouds/index.ts
@@ -1,0 +1,24 @@
+/*!
+ * Copyright 2022 Cognite AS
+ */
+
+export { intersectPointClouds } from './src/picking';
+
+export {
+  PotreePointShape,
+  PotreePointColorType,
+  PotreePointSizeType,
+  WellKnownAsprsPointClassCodes
+} from './src/types';
+
+export { PointCloudNode } from './src/PointCloudNode';
+export { PointCloudManager } from './src/PointCloudManager';
+export { createPointCloudManager } from './src/createPointCloudManager';
+export { PointCloudMetadata } from './src/PointCloudMetadata';
+
+export { PointCloudBudget } from './src/PointCloudBudget';
+export { PointCloudIntersection } from './src/PointCloudIntersection';
+export { CognitePointCloudModel } from './src/CognitePointCloudModel';
+
+export { PotreeNodeWrapper } from './src/PotreeNodeWrapper';
+export { PotreeGroupWrapper } from './src/PotreeGroupWrapper';
