+++ viewer/packages/pointclouds/src/PointCloudMetadata.ts
