+++ viewer/packages/pointclouds/src/PointCloudNode.ts
