+++ viewer/packages/pointclouds/src/PotreeGroupWrapper.ts
@@ -3,7 +3,7 @@
  */
 
 import * as Potree from '@cognite/potree-core';
-import { LoadingState } from '@reveal/cad-geometry-loaders';
+import { LoadingState } from '@reveal/model-base';
 
 import { combineLatest, interval, Observable, of, pipe, Subject } from 'rxjs';
 import { delay, distinctUntilChanged, map, share, startWith, switchMap } from 'rxjs/operators';
