+++ viewer/packages/pointclouds/src/PotreeNodeWrapper.ts
