+++ viewer/packages/pointclouds/src/picking.ts
